Configure OSD: check if disk already added
It is not possible to identify if the disk
is already added in microceph as list-disks
action returns disks by device-id and user
can add disks in format /dev/sdb. So rerun
the add-osd action and looks for error
message if disk is already added. The action
add-osd in charm microceph is modified to
return status messages for each disk [1].

Fixes: LP#2065649

[1] https://github.com/canonical/charm-microceph/pull/79