Fix microceph and hypervisor issues
microceph: fix wrongly initialized micrceph deploy in resize
hypervisor: AddHypervisorUnitsStep relied on a wrong behaviour of the
hypervisor getting to active status which does not happen when there's
no ceph relation. In case of no ceph relation, the hypervisor stays in
waiting. It only went to active because of the way status were treated
in sunbeam charms. This has been fixed.