refactor cluster list, implement for MAAS
Original cluster was just fetching members from the cluster, displaying
the corresponding role, and whether or not the microcluster was active.
It only worked in local mode.

In a MAAS deployment, very few nodes have a clusterd. It does not make
sense to tie the cluster list to the microcluster members.

Now, displays the machines in controller and openstack-machines (MAAS
only) models, with the status of each unit corresponding to its role.

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>