[tls ca] Wait for traefik and keystone apps to settle
Once tls unit certs are updated, it takes time for
traefik and keystone apps to settle as the traefik
application has to be restarted and ingress url has
to be updated. The related applications have to pick
up the ingress-changed event and change relation
data in identity-service to update the endpoints.

Add a step to wait until traefik and keystone units
are active and idle.