[vault] Add sunbeam commands to initialize and unseal vault
Move to latest charm release of vault. This requires additional
commands to initalize and unseal vault.

Changes to realise above:
* Use vault channel 1.16/stable
* Add new commands
  sunbeam vault init <key shares> <key threshold>
  sunbeam vault unseal <key>
  sunbeam vault authorize-charm <root token>

User has to run uneal command <key threshold> number
of times to unseal the leader unit. And have to rerun
the same commands again to unseal all the rest of units
(non-leader). This is necessary because the vault unseal
operation on non-leader unit fails when the vault leader
is sealed.