Fix wrong base deployed
Previous code was calling libjuju juju helper to add machine, but it did
not support using bases, and the series parameter is not actually used
when performing RPC call. Extracted necessary code and calling facade
directory to always deploy with a base.

This is necessary because 1) juju 3.6 default base is Noble 2) for the
move to Noble, we might need to still deploy one of the VMs with Jammy

Signed-off-by: Guillaume Boutry <guillaume.boutry@canonical.com>