Do not monitor the apps that are in blocked state
DeployControlPlane step can be run multiple times
when user runs `boostrap`, `resize` command in local mode
and `deploy` command in maas mode. However there
could be scenario where secrets or vault feature is
enabled in between the runs and the vault is not
configured to useal. In those cases, the
DeployControlPlane step need not monitor vault.
Otherwise the command runs until timeout.

Check if the apps that are blocked during feature
enablement are in active state or not. If the apps
are not in active state, remove the apps from the
monitoring during DeployControlPlane Step.
Currently barbican, vault fall under such apps.