Fix missing node name in cluster list
In case of machines deployed before
migration from lxd based juju controller
to k8s based juju controller, the hostname
is missing in juju model status [1]. Because of
which the cluster list command is broken and
not showing node name.

As a workaround for bug [1], if hostname is not
available in model status compare dns-name with
the IP address in microcluster status. If it matches
update the node name and cluster status based on
microcluster record.

Fixes: #2095063

[1] https://github.com/juju/juju/issues/18641